Add gcc wrapper to autotools when using CC_FOR_BUILD.
Autotools will not pass cflags to the CC_FOR_BUILD compiler when testing if it works, so our --sysroot does not make it through. Instead we can use the same approach we used to workaround the same issue in meson - generate a .sh file with the value of --sysroot and -B inlined inside it.